+++ implicitmf/__init__.py
@@ -8,6 +8,7 @@ recommendations for implicit feedback datasets.
 """
 
 from .transform import Transformer
-from .validation import hold_out_entries, gridsearchCV, cross_val_folds
+from .validation import hold_out_entries, cross_val_folds
+from .tune import gridsearchCV
 from .preprocess import normalize_X
 from .postprocess import remove_subscribed_items
