+++ tests/test_validation.py
@@ -9,7 +9,7 @@ import pytest
 import numpy as np
 import pandas as pd
 from scipy.sparse import csr_matrix, issparse
-from implicitmf.validation import hold_out_entries, cross_val_folds, gridsearchCV
+from implicitmf.validation import hold_out_entries, cross_val_folds
 from implicit.als import AlternatingLeastSquares
 from _mock_data import sparse_array
 
@@ -67,15 +67,3 @@ def test_cross_val_folds_values():
     output = cross_val_folds(X, n)
     assert(issparse(output[0]['train']))
     assert(issparse(output[0]['test']))
-
-def test_gridsearchCV_output():
-    """
-    Check that output of gridsearchCV()
-    is a pandas dataframe.
-    """
-    als = AlternatingLeastSquares()
-    hyperparams = {
-        'regularization': [0.2,0.3]
-    }
-    output = gridsearchCV(base_model=als, X=sparse_array(), n_folds=2, hyperparams=hyperparams)
-    assert isinstance(output, pd.DataFrame)
